backend/my_flask_app/requirements.txt
flask
flask-cors
mysql-connector-python
pycryptodome
ultralytics
opencv-python-headless
pillow
